+++ redux/src/main/renderer/index.js
@@ -6,7 +6,7 @@ import { Router, Route, IndexRoute } from 'react-router'
 import AccountInfo from './containers/AccountInfoContainer'
 import FavoritesContainer from './containers/FavoritesContainer'
 import MyTimelineContainer from './containers/MyTimelineContainer'
-import App from './containers/App'
+import App from './containers/AppContainer'
 import Following from './containers/FollowingContainer'
 import HomeContainer from './containers/HomeContainer.js'
 import configureStore from './store/configureStore'
